+++ tests/CommonItemsTests.vcxproj.filters
@@ -30,9 +30,6 @@
     <ClCompile Include="ColorTests.cpp">
       <Filter>tests</Filter>
     </ClCompile>
-    <ClCompile Include="..\newParser.cpp">
-      <Filter>commonItems</Filter>
-    </ClCompile>
     <ClCompile Include="..\Log.cpp">
       <Filter>commonItems</Filter>
     </ClCompile>
@@ -54,15 +51,21 @@
     <ClCompile Include="iconvliteTests.cpp">
       <Filter>tests</Filter>
     </ClCompile>
+    <ClCompile Include="LogTests.cpp">
+      <Filter>tests</Filter>
+    </ClCompile>
+    <ClCompile Include="..\Parser.cpp">
+      <Filter>commonItems</Filter>
+    </ClCompile>
+    <ClCompile Include="..\newParser.cpp">
+      <Filter>commonItems</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="..\CardinalToOrdinal.h">
       <Filter>commonItems</Filter>
     </ClInclude>
     <ClInclude Include="..\Color.h">
-      <Filter>commonItems</Filter>
-    </ClInclude>
-    <ClInclude Include="..\newParser.h">
       <Filter>commonItems</Filter>
     </ClInclude>
     <ClInclude Include="..\Log.h">
@@ -78,6 +81,12 @@
       <Filter>commonItems</Filter>
     </ClInclude>
     <ClInclude Include="..\Date.h">
+      <Filter>commonItems</Filter>
+    </ClInclude>
+    <ClInclude Include="..\Parser.h">
+      <Filter>commonItems</Filter>
+    </ClInclude>
+    <ClInclude Include="..\newParser.h">
       <Filter>commonItems</Filter>
     </ClInclude>
   </ItemGroup>
